add: feature pagination and search

+++ client/src/stores/counter.js
@@ -1,14 +1,16 @@
 import { defineStore } from 'pinia'
 import axios from 'axios'
-
+import Swal from 'sweetalert2'
 const baseUrl = "http://localhost:3000"
 
 export const useCounterStore = defineStore('counter', {
   state: () => ({
+    totalPage: 0,
     orderNumber: '',
     email: '',
     status: '',
     totalMovies: 0,
+    pagination: 0,
     movies: [],
     movie: [],
     isLogin: localStorage.access_token,
@@ -29,10 +31,22 @@ export const useCounterStore = defineStore('counter', {
           // console.log(response);
           this.router.push('/')
           localStorage.access_token = response.data.access_token
+          Swal.fire({
+            position: 'top-end',
+            icon: 'success',
+            title: 'You have successfully registered and logged in',
+            showConfirmButton: false,
+            timer: 2500
+          })
           // this.isLogin = response.data.access_token
         })
         .catch((error) => {
-          console.log(error);
+          // console.log(error);
+          Swal.fire({
+            icon: 'error',
+            title: 'Oops...',
+            text: error.response.data.message
+          })
         })
     },
     login(input) {
@@ -46,13 +60,26 @@ export const useCounterStore = defineStore('counter', {
       })
         .then((response) => {
           // console.log(response.data.access_token);
+          Swal.fire({
+            position: 'top-end',
+            icon: 'success',
+            title: 'You have successfully logged in',
+            showConfirmButton: false,
+            timer: 2500
+          })
           localStorage.access_token = response.data.access_token
           this.isLogin = response.data.access_token
           this.router.push('/')
           this.movie = []
+
         })
         .catch((error) => {
           console.log(error);
+          Swal.fire({
+            icon: 'error',
+            title: 'Oops...',
+            text: error.response.data.message
+          })
         })
     },
     googleLogin(data) {
@@ -65,9 +92,16 @@ export const useCounterStore = defineStore('counter', {
         },
       })
         .then((response) => {
+          Swal.fire({
+            position: 'top-end',
+            icon: 'success',
+            title: 'You have successfully logged in',
+            showConfirmButton: false,
+            timer: 2500
+          })
           console.log(response);
           localStorage.access_token = response.data.access_token;
-          this.isLogin = true;
+          this.isLogin = response.data.access_token
           this.router.push("/");
         })
         .catch(error => {
@@ -80,21 +114,40 @@ export const useCounterStore = defineStore('counter', {
         })
     },
     logout() {
-      localStorage.clear()
-      this.router.push('/')
+      Swal.fire({
+        title: 'Are you sure?',
+        text: "You won't be able to revert this!",
+        icon: 'warning',
+        showCancelButton: true,
+        confirmButtonColor: '#3085d6',
+        cancelButtonColor: '#d33',
+        confirmButtonText: 'Logout!'
+      }).then((result) => {
+        if (result.isConfirmed) {
+          localStorage.clear()
+          this.isLogin = ''
+          this.router.push('/')
+          Swal.fire(
+            'you have successfully logged out',
+          )
+        }
+      })
     },
-    getMovies() {
+    getMovies(params) {
       axios({
         method: 'GET',
         url: `${baseUrl}/movies`,
+        params: params,
         headers: {
-          'X-RapidAPI-Key': 'a19ac7857cmsh9fb1c28537353eep13be81jsnc2cf2963c3ac',
+          'X-RapidAPI-Key': '2e49e0f5b0msh29f18cffd1cf153p1e2596jsnb0ce53987731',
           'X-RapidAPI-Host': 'imdb-top-100-movies.p.rapidapi.com'
         }
       })
         .then((response) => {
           console.log(response);
           this.movies = response.data.data
+          this.totalMovies = response.data.data.length
+          this.totalPage = response.data.totalPage
         })
         .catch((error) => {
           console.log(error);
@@ -105,7 +158,7 @@ export const useCounterStore = defineStore('counter', {
         method: 'GET',
         url: `${baseUrl}/movie/${id}`,
         headers: {
-          'X-RapidAPI-Key': 'a19ac7857cmsh9fb1c28537353eep13be81jsnc2cf2963c3ac',
+          'X-RapidAPI-Key': '2e49e0f5b0msh29f18cffd1cf153p1e2596jsnb0ce53987731',
           'X-RapidAPI-Host': 'imdb-top-100-movies.p.rapidapi.com'
         }
       })
